fix(activities): the reading fires on every door, and a refused one keeps the transcript
Codex review of this branch, three findings, all real and all fixed.

IT ONLY FIRED ON ONE OF THREE DOORS. The hook sat in LogActivity, which opens
its own transaction. LogActivityTx — the shared-transaction variant driven by
POST /v1/activities and by the extension core — called the write body directly
and never offered a reading at all. So a transcript posted over REST was stored
and silently never read: the exact silence this feature exists to end,
reintroduced on two of the three doors it was meant to cover. Both entry points
now route through logActivityAndReadTranscript, and a fitness test reads
activity.go to hold that: logActivityInTx may have exactly one caller.

THE REST TRANSPORT HAD NO SEAM AT ALL. WithTranscriptRead wired the enqueue
onto the tool registry and nothing else, so even after the above, the handlers'
store carried a nil enqueue and skipped in silence. Handlers gained
WithTranscriptEnqueue and the option sets it. A new integration test posts a
transcript to POST /v1/activities and then asks for its reading — 200, with
nobody pressing anything.

A TOO-LONG TRANSCRIPT DESTROYED THE ACTIVITY. The reading runs inside the
activity's own transaction, so its error rolls the write back. That is right
for a database fault and wrong for WithinReadingBounds, which refuses past 600
lines / 60,000 characters while an activity body may be 256 KiB. A perfectly
legal activity vanished, and the caller was told its transcript was too long —
for a reading they never asked for. TranscriptTooLongError now skips the
reading and keeps the write. The explicit door keeps its 422, because a call
that asks for a reading and nothing else deserves the refusal.

One integration test changed premise rather than expectation.
TestATranscriptNobodyHasReadIsNotFoundRatherThanEmpty logged a transcript and
asserted 404. A transcript no longer has that state — landing reads it. The
property is still worth holding, so it now uses a plain meeting note, under its
own name, and the transcript case is covered by the new REST test above.

Also: activity.go crossed the 500-line cap, so the reads moved to
activityread.go. That exposed replayedActivity to the restricted-reader gate —
it had been passing under a sibling function's marker rather than its own. It
is safe and now says so in the ratification list: it reads the idempotency key
unscoped ON PURPOSE, because a scoped lookup would report a taken key as free,
and it discloses nothing — the id goes straight to readActivity, whose gate
turns a scope miss into the same 409 the unique index gives.
