+++ backend/restrictedreaders_test.go
@@ -83,6 +83,7 @@ var literalMarkers = []*regexp.Regexp{
 // another file — or that reach held rows on purpose. Each names the cost.
 var restrictedReadersAdmitted = gatekit.Waive(map[string]string{
 	"internal/compose/audiencerescope.go":         "the audience-change consumer reads the thread key (content by the activity policy) and the capture owner of the ONE activity whose audience just moved — deliberately, as a system principal, because both are exactly what narrowing the derived models needs, to NARROW what other readers may see — excluding a held row here would leave a legal-hold conversation's derived signals workspace-visible, the exact disclosure the consumer exists to remove. The cost is that a held activity's thread key and owner id reach this system principal",
+	"internal/modules/activities/activity.go":     "replayedActivity resolves the (source_system, source_id) idempotency key and deliberately reads it UNSCOPED, because the question it asks is whether the key is taken — a scoped lookup would answer 'free' for a key held by a row out of scope, and the insert behind it would then fail on the unique index instead. It discloses nothing on its own: the id it finds is handed straight to readActivity (activityread.go), which carries the row scope, and a scope miss is turned into the same 409 the unique-index race returns — the key is taken, the record is not shown. The cost is that this reader depends on readActivity keeping its gate, in a different file since the write and read halves were split at the 500-line cap",
 	"internal/modules/activities/capturelabel.go": "the classify backlog reads subject and body through ClassifyBacklogPredicate (pipelinefacts.go), which filters archived_at IS NULL and so excludes held rows by the restricted-implies-archived CHECK — the cost is that the predicate's archived filter now carries this obligation as well as its own",
 })
